Fix profile details not viewable when logged out
